requirements.txt
# ── Deep Learning ─────────────────────────────────────────────────────────────
tensorflow==2.13.0          # Core DL framework; VGG16 is in tf.keras.applications
keras==2.13.1

# ── Data & Image Processing ───────────────────────────────────────────────────
#pandas==2.3.0
numpy==1.24.3
#matplotlib==3.10.8
#seaborn==0.13.2
Pillow==10.0.1              # Image I/O (PIL)
opencv-python-headless==4.8.0.76   # headless: no GUI deps (safe for Docker)
scikit-learn==1.3.0         # train_test_split, metrics

# ── MLOps: Experiment Tracking ────────────────────────────────────────────────
mlflow==2.7.1               # Experiment tracking, model registry, artifact store
dagshub                     # Remote MLflow + DVC tracking server (free tier)

# ── MLOps: Data Versioning ────────────────────────────────────────────────────
dvc==3.25.0                 # Data Version Control
dvc[s3]                     # Optional: S3 remote storage for DVC artifacts

# ── API & Web Serving ─────────────────────────────────────────────────────────
fastapi==0.103.1            # High-performance async REST API
uvicorn[standard]==0.23.2   # ASGI server for FastAPI
python-multipart==0.0.6     # Required for FastAPI file uploads
jinja2==3.1.2               # HTML templating for FastAPI
aiofiles==23.2.1            # Async static file serving

# ── Configuration & Utilities ─────────────────────────────────────────────────
PyYAML==6.0.1               # Read config.yaml, params.yaml, schema.yaml
python-box==7.1.1           # Dot-notation access for dicts (box['a'] → box.a)
ensure==1.0.3               # Type validation decorators
pathlib                     # Built-in, listed for explicitness
tqdm==4.66.1                # Progress bars for data download

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==7.4.2
pytest-cov==4.1.0           # Coverage reports

# ── Code Quality ──────────────────────────────────────────────────────────────
flake8==6.1.0               # Linting
black==23.9.1               # Auto-formatting
isort==5.12.0               # Import sorting

# ── Notebooks (research/ only — not imported by production code) ──────────────
#jupyter==1.0.0
#ipykernel==6.25.2

-e .